+++ tests/DependencyInjection/ConfigurationTest.php
@@ -20,6 +20,7 @@ final class ConfigurationTest extends TestCase
     {
         $expectedBundleDefaultConfig = [
             'register_error_listener' => true,
+            'transport_factory' => 'Sentry\\Transport\\TransportFactoryInterface',
             'options' => [
                 'integrations' => [],
                 'prefixes' => array_merge(['%kernel.project_dir%'], array_filter(explode(\PATH_SEPARATOR, get_include_path() ?: ''))),